requirements.txt
textual
requests
dohq-artifactory
pyyaml
